.pkms/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pkms"
version = "0.3.0"
description = "Personal Knowledge Management System with hybrid search and semantic chunking"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "PKMS Contributors"}
]
keywords = ["knowledge-management", "zettelkasten", "search", "embeddings", "markdown"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "python-frontmatter>=1.0.0",
    "python-slugify>=8.0.0",
    "python-ulid>=2.0.0",
    "langdetect>=1.0.9",
    "whoosh>=2.7.4",
    "numpy>=1.24.0",
    "ollama>=0.1.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

performance = [
    "xxhash>=3.0.0",
    "tiktoken>=0.5.0",
]

validation = [
    "check-jsonschema>=0.27.0",
]

all = [
    "pkms[dev,performance,validation]",
]

[project.scripts]
pkms-ingest = "tools.ingest:main"
pkms-chunk = "tools.chunk:main"
pkms-link = "tools.link:main"
pkms-embed = "tools.embed_index:main"
pkms-relevance = "tools.relevance:main"
pkms-archive = "tools.archive:main"
pkms-synth = "tools.synth:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["lib*", "models*", "tools*"]
exclude = ["tests*", "docs*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "frontmatter.*",
    "langdetect.*",
    "whoosh.*",
    "ollama.*",
    "xxhash.*",
    "tiktoken.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["pkms"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
